Create Simple string length measurer
Code Description
The code is a simple C program that reads a user's name as input, calculates the length of the entered string manually, and also calculates the length using the strlen function from the <string.h> library. The program then prints both the lengths and the entered name.

Code Breakdown
Includes and Constants:

The program includes the stdio.h library for standard input and output functions and the string.h library for string manipulation functions.
A constant integer num is defined with a value of 10, which specifies the maximum length of the name that can be entered.
Main Function:

The main function starts the execution of the program.
An array name of size num is declared to store the entered name.
An integer charCount is initialized to 0 to count the characters in the string.
A char variable placeholder is declared but not used in the program.
User Input:

The program prompts the user to enter their name using printf.
scanf is used to read the entered name into the name array. Note that scanf("%s", name); is used correctly to read the string, but scanf("%s", &name); should be scanf("%s", name); because name already acts as a pointer to the first element of the array.
Manual String Length Calculation:

A while loop iterates through the name array until it encounters the null terminator '\0', incrementing charCount to calculate the length of the string manually.
Output:

The program prints the manually calculated length of the string.
It also prints the entered name using printf.
String Length Using strlen:

The strlen function from the string.h library is used to calculate the length of the string.
The program prints the length of the string calculated using strlen.